requirements.txt
openai
flask
PyYAML
rasa